+++ spec/braintree_web_drop_in_spec.rb
@@ -65,6 +65,44 @@ describe "Drop-in" do
     end
   end
 
+  describe "promise API" do
+    it "tokenizes a card" do
+      visit "http://#{HOSTNAME}:#{PORT}/promise.html"
+
+      click_option("card")
+      hosted_field_send_input("number", "4111111111111111")
+      hosted_field_send_input("expirationDate", "1019")
+      hosted_field_send_input("cvv", "123")
+
+      submit_pay
+
+      expect(find(".braintree-heading")).to have_content("Paying with")
+
+      # Drop-in Details
+      expect(page).to have_content("Ending in ••11")
+
+      # Nonce Details
+      expect(page).to have_content("CreditCard")
+      expect(page).to have_content("ending in 11")
+      expect(page).to have_content("Visa")
+    end
+
+    it "tokenizes PayPal", :paypal do
+      visit "http://#{HOSTNAME}:#{PORT}/promise.html"
+
+      click_option("paypal")
+
+      open_popup_and_complete_login
+
+      submit_pay
+
+      expect(find(".braintree-heading")).to have_content("Paying with PayPal")
+
+      expect(page).to have_content("PayPalAccount")
+      expect(page).to have_content(ENV["PAYPAL_USERNAME"])
+    end
+  end
+
   describe "updateConfiguration" do
     it "updates PayPal configuration", :paypal do
       visit "http://#{HOSTNAME}:#{PORT}?showUpdatePayPalMenu=true"